fix: complete electronic invoicing pipeline for Hacienda CR v4.4
- Rewrite FirmaDigitalServiceImpl to XAdES-BES (was plain XMLDSig)
- Fix XML serialization: INDENT=no to preserve signature integrity
- Add ProveedorSistemas element at position 2 per v4.4 schema
- Extract NumeroConsecutivo from Clave (pos 21-40) to ensure consistency
- Guard OtrasSenas minLength=5 in Emisor and skip Receptor/Ubicacion when invalid
- Fix 202 response: treat as ENVIADO (async processing), not RECHAZADO
- Fix ind-estado parsing: Hacienda returns hyphen-key, not camelCase
- Set indicadorEstado in RespuestaHaciendaDTO builder from parsed response
- Add retry loop (5x30s) in procesarFacturaAsync to poll final state
- Decode base64 MensajeHacienda in comprobantes UI modal

+++ src/main/java/api/astro/whats_orders_manager/modules/facturacion/electronica/controller/ConfiguracionHaciendaController.java
@@ -39,7 +39,7 @@ public class ConfiguracionHaciendaController {
     public ResponseEntity<ConfiguracionHaciendaDTO> obtenerConfiguracionActiva(
             @PathVariable Integer empresaId) {
         log.info("Obteniendo configuración activa para empresa {}", empresaId);
-        return configuracionService.obtenerConfiguracionActiva(empresaId.longValue())
+        return configuracionService.obtenerConfiguracionActiva(empresaId)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
     }
@@ -52,7 +52,7 @@ public class ConfiguracionHaciendaController {
     public ResponseEntity<List<ConfiguracionHaciendaDTO>> listarConfiguraciones(
             @PathVariable Integer empresaId) {
         log.info("Listando configuraciones de Hacienda para empresa {}", empresaId);
-        List<ConfiguracionHaciendaDTO> configuraciones = configuracionService.listarPorEmpresa(empresaId.longValue());
+        List<ConfiguracionHaciendaDTO> configuraciones = configuracionService.listarPorEmpresa(empresaId);
         return ResponseEntity.ok(configuraciones);
     }
     
@@ -65,7 +65,7 @@ public class ConfiguracionHaciendaController {
             @PathVariable Integer empresaId,
             @PathVariable AmbienteHacienda ambiente) {
         log.info("Obteniendo configuración {} para empresa {}", ambiente, empresaId);
-        ConfiguracionHacienda config = configuracionService.obtenerPorEmpresaYAmbiente(empresaId.longValue(), ambiente)
+        ConfiguracionHacienda config = configuracionService.obtenerPorEmpresaYAmbiente(empresaId, ambiente)
                 .orElse(null);
         if (config == null) {
             return ResponseEntity.notFound().build();
